New draft of the user manual. Issue 0.9.7, draft A. Update to correct things left unfinished in manual for the 0.9.6 release of ArgoUML. The reference part of the manual (chapters 7 through 22) is now completed. In addition chapter 1 (introduction), tutorial chapter 2 (UML based &ooad;) are largely complete.</para>
The comments on the 0.9.6 issue should have referred to chapter 21 as covering
deployment diagrams (partially completed) and chapter 22 as covering built in
datatypes, classes, interfaces and stereotypes. The comments should also have
indicated that chapter 20 (activity diagrams) was complete.

Updated to complete chapter 21 (deployment diagrams). Bug fixes of navigation
between interfaces and classes and extensions to the pop-up menu for classes
have been noted, and the documentation revised accordingly.


git-svn-id: http://argouml.stage.tigris.org/svn/argouml/trunk@1422 a161b567-7d1e-0410-9ef9-912c70fedb3f