THESE FILES ARE CONTRIBUTED BY NOVOSOFT. BELOW IS AN EXCERPT FROM A MAIL RECEIVED BY MARKUS KLINK (10/30/2002):
You are granted permission to publish the document
and derived versions of it as part of argouml developer
cook book on argouml site. You also can make changes
to document to reflect changes in nsuml since time document
is written. We could later integrate changes you had made
back to nsuml documentation.

In case of changes that fall ouside of formating and
splitting to parts, please indicate it somewhere document.


git-svn-id: http://argouml.stage.tigris.org/svn/argouml/trunk@2930 a161b567-7d1e-0410-9ef9-912c70fedb3f