The next release will be the stable 0.22 release.
Please help in testing the 0.22.beta3.

If you think the 0.22.beta3 release is not good enough for the stable release,
write this in the DEFECT describing the release together with the low-risk
solution to the problem (in a branch or using some other mean). Then notify
the release responsible.

If we decide to include the fix we will also introduce a beta4 release.


git-svn-id: http://argouml.stage.tigris.org/svn/argouml/trunk@10965 a161b567-7d1e-0410-9ef9-912c70fedb3f