Corrected a potential problem located by FindBugs :
File: PriorityNode.java, Line: 70, Type: LI_LAZY_INIT_UPDATE_STATIC, Priority: High, Category: CORRECTNESS

Incorrect lazy initialization and update of static field

This method contains an unsynchronized lazy initialization of a static field. After the field is set, the object stored into that location is further accessed. The setting of the field is visible to other threads as soon as it is set. If the futher accesses in the method that set the field serve to initialize the object, then you have a very serious multithreading bug, unless something else prevents any other thread from accessing the stored object until it is fully initialized. 

git-svn-id: http://argouml.stage.tigris.org/svn/argouml/trunk@17358 a161b567-7d1e-0410-9ef9-912c70fedb3f